# openapi.yaml
{
  "openapi": "3.0.1",
  "info": {
    "title": "Biere Catalog API",
    "description": "API for managing beers, breweries, styles, and countries.",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/v1/users/login": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Login",
        "description": "Authenticates a user and returns a token.",
        "operationId": "getToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponseDTO"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/beers": {
      "get": {
        "tags": [
          "Beers"
        ],
        "summary": "Get all beers (JSON)",
        "description": "Retrieves a paginated list of beers in JSON format.",
        "operationId": "getBeersCsv",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Pagination information",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved list",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Beers"
        ],
        "summary": "Register a new beer",
        "description": "Creates a new beer record.",
        "operationId": "register_2",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BeerRegistrationDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Beer created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiIndividualResponseDTOBeerResponseDTO"
                }
              },
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiIndividualResponseDTOBeerResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiIndividualResponseDTOBeerResponseDTO"
                }
              },
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiIndividualResponseDTOBeerResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/v1/beers/{beerId}": {
      "get": {
        "tags": [
          "Beers"
        ],
        "summary": "Get a specific beer",
        "description": "Retrieves details of a specific beer by ID.",
        "operationId": "getSpecificBeer",
        "parameters": [
          {
            "name": "beerId",
            "in": "path",
            "description": "ID of the beer to be retrieved",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "404": {
            "description": "Beer not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiIndividualResponseDTOBeerResponseDTO"
                }
              }
            }
          },
          "200": {
            "description": "Successfully retrieved beer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiIndividualResponseDTOBeerResponseDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Beers"
        ],
        "summary": "Delete a beer",
        "description": "Deletes a beer by ID.",
        "operationId": "deleteBeer",
        "parameters": [
          {
            "name": "beerId",
            "in": "path",
            "description": "ID of the beer to be deleted",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "404": {
            "description": "Beer not found"
          },
          "204": {
            "description": "Beer deleted successfully"
          }
        }
      },
      "patch": {
        "tags": [
          "Beers"
        ],
        "summary": "Update a beer",
        "description": "Updates an existing beer by ID.",
        "operationId": "updateBeer",
        "parameters": [
          {
            "name": "beerId",
            "in": "path",
            "description": "ID of the beer to be updated",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BeerUpdateRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Beer not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiIndividualResponseDTOBeerResponseDTO"
                }
              }
            }
          },
          "200": {
            "description": "Beer updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiIndividualResponseDTOBeerResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Links": {
        "type": "object",
        "additionalProperties": {
          "$ref": "#/components/schemas/Link"
        }
      },
...
      "TokenResponseDTO": {
        "required": [
          "access_token",
          "refresh_token"
        ],
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "description": "JWT access token",
            "example": "eyJhbGciOiJIUzI1NiJ9..."
          },
          "refresh_token": {
            "type": "string",
            "description": "JWT refresh token",
            "example": "eyJhbGciOiJIUzI1NiJ9..."
          }
        },
        "description": "DTO for authentication response containing tokens"
      },
      "TokenRequestDTO": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "john.doe@example.com"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "format": "password",
            "example": "securePassword123"
          }
        },
        "description": "DTO for authentication request"
      },
...
      "BeerRegistrationDTO": {
        "required": [
          "alcohol_content",
          "brewery_id",
          "name",
          "style_id",
          "year"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string",
            "description": "Name of the beer",
            "example": "Heineken"
          },
          "alcohol_content": {
            "maximum": 100,
            "minimum": 0.0,
            "type": "number",
            "description": "Alcohol content percentage",
            "format": "float",
            "example": 5.0
          },
          "brewery_id": {
            "type": "integer",
            "description": "ID of the brewery",
            "format": "int64",
            "example": 1
          },
          "style_id": {
            "type": "integer",
            "description": "ID of the beer style",
            "format": "int64",
            "example": 1
          },
          "year": {
            "type": "integer",
            "description": "Year of production",
            "format": "int64",
            "example": 2023
          }
        },
        "description": "DTO for beer registration request"
      },
      "ApiIndividualResponseDTOBeerResponseDTO": {
        "required": [
          "data"
        ],
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/BeerResponseDTO"
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        },
        "description": "Generic wrapper for a single resource response"
      },
      "BeerResponseDTO": {
        "required": [
          "alcohol_content",
          "brewery",
          "country_name",
          "id",
          "name",
          "style",
          "year"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique identifier of the beer",
            "format": "int64",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Name of the beer",
            "example": "Heineken"
          },
          "country_name": {
            "type": "string",
            "description": "Country of origin",
            "example": "Netherlands"
          },
          "alcohol_content": {
            "type": "number",
            "description": "Alcohol content percentage",
            "format": "float",
            "example": 5.0
          },
          "brewery": {
            "type": "string",
            "description": "Name of the brewery",
            "example": "Heineken Brewery"
          },
          "style": {
            "type": "string",
            "description": "Style of the beer",
            "example": "Lager"
          },
          "year": {
            "type": "integer",
            "description": "Year of production",
            "format": "int64",
            "example": 2023
          }
        },
        "description": "DTO for beer response data"
      },
      "BeerUpdateRequestDTO": {
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 100,
            "type": "string",
            "description": "New name of the beer",
            "example": "Heineken Silver"
          },
          "alcohol_content": {
            "maximum": 100.0,
            "minimum": 0.0,
            "type": "number",
            "description": "New alcohol content percentage",
            "format": "float",
            "example": 4.0
          },
          "brewery_id": {
            "type": "integer",
            "description": "New ID of the brewery",
            "format": "int64",
            "example": 1
          },
          "style_id": {
            "type": "integer",
            "description": "New ID of the beer style",
            "format": "int64",
            "example": 1
          },
          "year": {
            "type": "integer",
            "description": "New year of production",
            "format": "int64",
            "example": 2024
          }
        },
        "description": "DTO for beer update request (partial update)"
      },
...
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "Link": {
        "type": "object",
        "properties": {
          "href": {
            "type": "string"
          },
          "hreflang": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "deprecation": {
            "type": "string"
          },
          "profile": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "templated": {
            "type": "boolean"
          }
        }
      }
    }
  }
}